+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Flat map for nested properties.
 - Mapped observable list.
+- Filtered observable list.
+- Sorted observable list.
 
 ### Changed
 - rename `map` to `mapBinding` to avoid conflicts with `List.map()`